Moved depsgraph_update to properties/object.py as more related.
