+++ src/hdusd/engine/handlers.py
@@ -38,10 +38,10 @@ def on_load_post(*args):
 
 @bpy.app.handlers.persistent
 def on_depsgraph_update_post(scene, depsgraph):
-    from ..properties import usd_list
+    from ..properties import object
     from ..usd_nodes import node_tree
 
-    usd_list.depsgraph_update(depsgraph)
+    object.depsgraph_update(depsgraph)
     node_tree.depsgraph_update(depsgraph)
